+++ packages/client/src/pages/admin/overview.user.vue
@@ -1,12 +1,12 @@
 <template>
-<div :class="[$style.root, { yellow: user.isSilenced, red: user.isSuspended, gray: false }]">
+<MkA :class="[$style.root]" :to="`/user-info/${user.id}`">
 	<MkAvatar class="avatar" :user="user" :disable-link="true" :show-indicator="true"/>
 	<div class="body">
 		<span class="name"><MkUserName class="name" :user="user"/></span>
 		<span class="sub"><span class="acct _monospace">@{{ acct(user) }}</span></span>
 	</div>
-	<MkMiniChart v-if="chartValues" class="chart" :src="chartValues"/>
-</div>
+	<MkMiniChart v-if="chart" class="chart" :src="chart.inc"/>
+</MkA>
 </template>
 
 <script lang="ts" setup>
@@ -19,12 +19,10 @@ const props = defineProps<{
 	user: misskey.entities.User;
 }>();
 
-let chartValues = $ref<number[] | null>(null);
+const chart = $ref(null);
 
-os.apiGet('charts/user/notes', { userId: props.user.id, limit: 16 + 1, span: 'day' }).then(res => {
-	// 今日のぶんの値はまだ途中の値であり、それも含めると大抵の場合前日よりも下降しているようなグラフになってしまうため今日は弾く
-	res.inc.splice(0, 1);
-	chartValues = res.inc;
+os.apiGet('charts/user/notes', { userId: props.user.id, limit: 16, span: 'day' }).then(res => {
+	chart = res;
 });
 </script>
 
@@ -35,9 +33,6 @@ os.apiGet('charts/user/notes', { userId: props.user.id, limit: 16 + 1, span: 'da
 
 	display: flex;
 	align-items: center;
-	padding: 16px;
-	background: var(--panel);
-	border-radius: 8px;
 
 	> :global(.avatar) {
 		display: block;
@@ -76,24 +71,6 @@ os.apiGet('charts/user/notes', { userId: props.user.id, limit: 16 + 1, span: 'da
 
 	> :global(.chart) {
 		height: 30px;
-	}
-
-	&:global(.yellow) {
-		--c: rgb(255 196 0 / 15%);
-		background-image: linear-gradient(45deg, var(--c) 16.67%, transparent 16.67%, transparent 50%, var(--c) 50%, var(--c) 66.67%, transparent 66.67%, transparent 100%);
-		background-size: 16px 16px;
-	}
-
-	&:global(.red) {
-		--c: rgb(255 0 0 / 15%);
-		background-image: linear-gradient(45deg, var(--c) 16.67%, transparent 16.67%, transparent 50%, var(--c) 50%, var(--c) 66.67%, transparent 66.67%, transparent 100%);
-		background-size: 16px 16px;
-	}
-
-	&:global(.gray) {
-		--c: var(--bg);
-		background-image: linear-gradient(45deg, var(--c) 16.67%, transparent 16.67%, transparent 50%, var(--c) 50%, var(--c) 66.67%, transparent 66.67%, transparent 100%);
-		background-size: 16px 16px;
 	}
 }
 </style>